refactor(db): extract message action repository
